Remove old storage and registry code. Update app data fetching.
