+++ src/utils/color.ts
@@ -1,4 +1,4 @@
 import { HEX_COLOR_REGEXP, RGB_COLOR_REGEXP, HSL_COLOR_REGEXP } from '../constants'
 
 export const isValidColor = (color: string) =>
-  Boolean(color.match(HEX_COLOR_REGEXP) || color.match(RGB_COLOR_REGEXP) || color.match(HSL_COLOR_REGEXP))
+  !!color && Boolean(color.match(HEX_COLOR_REGEXP) || color.match(RGB_COLOR_REGEXP) || color.match(HSL_COLOR_REGEXP))
